refactor display filter, #244
