+++ client/main.js
@@ -2,8 +2,6 @@ var page = require('page');
 
 require('./middleware/initial-render')(page);
 
-// define the state of the filter pannel, open or not
-require('./middleware/filter-state')(page);
 // Client routes
 require('./routes/home')(page);
 require('./routes/search')(page);
