--- client/middleware/filter-state.js
@@ -1,7 +1,0 @@
-var filterState = require('../lib/filter-state');
-module.exports = function (page) {
-  page(function (ctx, next) {
-    ctx.isFilterOpen = filterState.isFilterOpen;
-    next();
-  });
-};
