+++ client/routes/search.js
@@ -5,13 +5,13 @@ var Templates = require('../templates');
 var createQueryParams = require('../../lib/query-params/query-params');
 var getData = require('../lib/get-data.js');
 var toJsonUrl = require('../lib/to-json-url');
-var filterState = require('../lib/filter-state');
+var displayFilters = require('../lib/display-filters.js');
 var filterResults = require('../lib/filter-results');
 var page = require('page');
 var searchResultsToTemplateData = require('../../lib/transforms/search-results-to-template-data');
 var searchListener = require('../lib/search-listener');
 var Snackbar = require('snackbarlightjs');
-var toggleClass = require('../js-helpers/toggleClass');
+var filterState = require('../lib/filter-state.js');
 var i = 0;
 
 module.exports = function (page) {
@@ -88,21 +88,25 @@ function render (ctx, next) {
 function listeners (ctx, next) {
   searchListener();
 
-  var toggleElements = function () {
-    var searchresults = document.querySelector('.searchresults');
-    searchresults.className = toggleClass(searchresults.className, 'searchresults--filtersactive');
-
-    var filtercolumn = document.querySelector('.filtercolumn');
-    filtercolumn.className = toggleClass(filtercolumn.className, 'filtercolumn--filtersactive');
-
-    var controlFilters = document.querySelector('.control--filters');
-    controlFilters.className = toggleClass(controlFilters.className, 'control--active');
-  };
-
-  var controlButtons = document.getElementsByClassName('control__button');
-  for (i = 0; i < controlButtons.length; i++) {
-    controlButtons[i].addEventListener('click', toggleElements);
+  // hide the filter button
+  var filterButton = document.querySelector('button.filterpanel__button');
+  if (filterButton) {
+    filterButton.style.display = 'none';
   }
+
+  // display the button on the filter which toggle the filters
+  var toggleFilterButton = document.getElementById('fb');
+  if (toggleFilterButton) {
+    toggleFilterButton.classList.remove('hidden');
+    toggleFilterButton.classList.add('control__button');
+  }
+
+  // add click event listner on fb button to toggle the filter
+  displayFilters(filterState.isFilterOpen);
+  toggleFilterButton.addEventListener('click', function () {
+    filterState.isFilterOpen = !filterState.isFilterOpen;
+    displayFilters(filterState.isFilterOpen);
+  });
 
   /**
   * Click to add/remove filters
@@ -136,22 +140,9 @@ function listeners (ctx, next) {
     });
   }
 
-  // $('.control--rpp select').on('change', function (e) {
-  //   filterResults(ctx, page);
-  // });
-
-  /**
-  * update filter status (open/close)
-  */
-  var filterButton = document.querySelector('#fb');
-  if (filterButton) {
-    filterButton.addEventListener('click', function () {
-      filterState.isFilterOpen = !filterState.isFilterOpen;
-    });
-  }
-
   initComp();
 
+  // analytics
   const onResultClick = (e) => {
     const id = e.currentTarget.getAttribute('href').split('/').pop();
